+++ internal/js_parser/json_parser.go
@@ -9,10 +9,10 @@ import (
 )
 
 type jsonParser struct {
-	log                 logger.Log
-	source              logger.Source
-	lexer               js_lexer.Lexer
-	allowTrailingCommas bool
+	log     logger.Log
+	source  logger.Source
+	lexer   js_lexer.Lexer
+	options JSONOptions
 }
 
 func (p *jsonParser) parseMaybeTrailingComma(closeToken js_lexer.T) bool {
@@ -20,7 +20,7 @@ func (p *jsonParser) parseMaybeTrailingComma(closeToken js_lexer.T) bool {
 	p.lexer.Expect(js_lexer.TComma)
 
 	if p.lexer.Token == closeToken {
-		if !p.allowTrailingCommas {
+		if !p.options.AllowTrailingCommas {
 			p.log.AddRangeError(&p.source, commaRange, "JSON does not support trailing commas")
 		}
 		return false
@@ -119,7 +119,7 @@ func (p *jsonParser) parseExpr() js_ast.Expr {
 			// Warn about duplicate keys
 			keyText := js_lexer.UTF16ToString(keyString)
 			if duplicates[keyText] {
-				p.log.AddRangeWarning(&p.source, keyRange, fmt.Sprintf("Duplicate key: %q", keyText))
+				p.log.AddRangeWarning(&p.source, keyRange, fmt.Sprintf("Duplicate key %q in object literal", keyText))
 			} else {
 				duplicates[keyText] = true
 			}
@@ -167,10 +167,10 @@ func ParseJSON(log logger.Log, source logger.Source, options JSONOptions) (resul
 	}()
 
 	p := &jsonParser{
-		log:                 log,
-		source:              source,
-		lexer:               js_lexer.NewLexerJSON(log, source, options.AllowComments),
-		allowTrailingCommas: options.AllowTrailingCommas,
+		log:     log,
+		source:  source,
+		options: options,
+		lexer:   js_lexer.NewLexerJSON(log, source, options.AllowComments),
 	}
 
 	result = p.parseExpr()